+++ src/Persistence/IdentifiablePostgresPersistence.cs
@@ -21,7 +21,8 @@ namespace PipServices3.Postgres.Persistence
     /// 
     /// ### Configuration parameters ###
     /// 
-    /// - collection:                  (optional) PostgreSQL collection name
+    /// - table:                     (optional) PostgreSQL table name
+    /// - schema:                    (optional) PostgreSQL schema name
     /// 
     /// connection(s):
     /// - discovery_key:             (optional) a key to retrieve the connection from <a href="https://pip-services3-dotnet.github.io/pip-services3-components-dotnet/interface_pip_services_1_1_components_1_1_connect_1_1_i_discovery.html">IDiscovery</a>
@@ -98,11 +99,17 @@ namespace PipServices3.Postgres.Persistence
         where K : class
     {
         /// <summary>
+        /// Flag to turn on auto generation of object ids.
+        /// </summary>
+        protected bool _autoGenerateId = true;
+
+        /// <summary>
         /// Creates a new instance of the persistence component.
         /// </summary>
-        /// <param name="tableName">(optional) a collection name.</param>
-        public IdentifiablePostgresPersistence(string tableName)
-            : base(tableName)
+        /// <param name="tableName">(optional) a table name.</param>
+        /// <param name="schemaName">(optional) a schema name.</param>
+        public IdentifiablePostgresPersistence(string tableName = null, string schemaName = null)
+            : base(tableName, schemaName)
         { }
 
         /// <summary>
@@ -114,7 +121,7 @@ namespace PipServices3.Postgres.Persistence
         public virtual async Task<List<T>> GetListByIdsAsync(string correlationId, K[] ids)
         {
             var @params = GenerateParameters(ids);
-            var query = "SELECT * FROM " + QuoteIdentifier(_tableName) + " WHERE \"id\" IN (" + @params + ")";
+            var query = "SELECT * FROM " + QuotedTableName() + " WHERE \"id\" IN (" + @params + ")";
 
             var items = await ExecuteReaderAsync(query, ids);
 
@@ -132,7 +139,7 @@ namespace PipServices3.Postgres.Persistence
         public virtual async Task<T> GetOneByIdAsync(string correlationId, K id)
         {
             var @params = new[] { id };
-            var query = "SELECT * FROM " + QuoteIdentifier(_tableName) + " WHERE \"id\" = @Param1";
+            var query = "SELECT * FROM " + QuotedTableName() + " WHERE \"id\" = @Param1";
 
             var result = (await ExecuteReaderAsync(query, @params)).FirstOrDefault();
 
@@ -157,8 +164,9 @@ namespace PipServices3.Postgres.Persistence
         /// <returns>created item.</returns>
         public override async Task<T> CreateAsync(string correlationId, T item)
         {
-            if (item is IStringIdentifiable && item.Id == null)
-                ObjectWriter.SetProperty(item, nameof(item.Id), IdGenerator.NextLong());
+            // Assign unique id
+            if (item is IStringIdentifiable stringIdentifiable && stringIdentifiable.Id == null && _autoGenerateId)
+                stringIdentifiable.Id = IdGenerator.NextLong();
 
             return await base.CreateAsync(correlationId, item);
         }
@@ -174,13 +182,17 @@ namespace PipServices3.Postgres.Persistence
             if (item == null || item.Id == null)
                 return default;
 
+            // Assign unique id
+            if (item is IStringIdentifiable stringIdentifiable && stringIdentifiable.Id == null && _autoGenerateId)
+                stringIdentifiable.Id = IdGenerator.NextLong();
+
             var map = ConvertFromPublic(item);
             var columns = GenerateColumns(map);
             var @params = GenerateParameters(map);
             var setParams = GenerateSetParameters(map);
             var values = GenerateValues(map);
 
-            var query = "INSERT INTO " + QuoteIdentifier(_tableName) + " (" + columns + ")"
+            var query = "INSERT INTO " + QuotedTableName() + " (" + columns + ")"
                 + " VALUES (" + @params +")"
                 + " ON CONFLICT (\"id\") DO UPDATE SET " + setParams + " RETURNING *";
 
@@ -208,7 +220,7 @@ namespace PipServices3.Postgres.Persistence
             var values = GenerateValues(map);
             values.Add(item.Id);
 
-            var query = "UPDATE " + QuoteIdentifier(_tableName)
+            var query = "UPDATE " + QuotedTableName()
                 + " SET " + @params +" WHERE \"id\"=@Param" + values.Count + " RETURNING *";
 
             var result = (await ExecuteReaderAsync(query, values)).FirstOrDefault();
@@ -236,7 +248,7 @@ namespace PipServices3.Postgres.Persistence
             var values = GenerateValues(map);
             values.Add(id);
 
-            var query = "UPDATE " + QuoteIdentifier(_tableName)
+            var query = "UPDATE " + QuotedTableName()
                 + " SET " + @params + " WHERE \"id\" = @Param" + values.Count + " RETURNING *";
 
             var result = (await ExecuteReaderAsync(query, values)).FirstOrDefault();
@@ -257,7 +269,7 @@ namespace PipServices3.Postgres.Persistence
         {
             var values = new[] { id };
 
-            var query = "DELETE FROM " + QuoteIdentifier(_tableName) + " WHERE \"id\" = @Param1 RETURNING *";
+            var query = "DELETE FROM " + QuotedTableName() + " WHERE \"id\" = @Param1 RETURNING *";
 
             var result = (await ExecuteReaderAsync(query, values)).FirstOrDefault();
 
@@ -275,7 +287,7 @@ namespace PipServices3.Postgres.Persistence
         public virtual async Task DeleteByIdsAsync(string correlationId, K[] ids)
         {
             var @params = GenerateParameters(ids);
-            var query = "DELETE FROM " + QuoteIdentifier(_tableName) + " WHERE \"id\" IN (" + @params +")";
+            var query = "DELETE FROM " + QuotedTableName() + " WHERE \"id\" IN (" + @params +")";
 
             var result = await ExecuteNonQuery(query, ids);