+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> PostgreSQL components for .NET Changelog
 
+## <a name="3.3.0"></a> 3.3.0 (2021-06-15) 
+
+### Features
+* Added schemas to PostgresPersistence, IdentifiablePostgresPersistence, IdentifiableJsonPostgresPersistence
+* Added _autoGenerateId flag to IdentifiablePostgresPersistence
+
 ## <a name="3.2.0"></a> 3.2.0 (2021-06-10) 
 
 ### Features